Rename "domain" package to "store".
This name more accurately reflects the contents.
