+++ server/src/main/java/com/adobe/testing/s3mock/S3MockConfiguration.java
@@ -16,10 +16,10 @@
 
 package com.adobe.testing.s3mock;
 
-import com.adobe.testing.s3mock.domain.FileStore;
-import com.adobe.testing.s3mock.domain.KmsKeyStore;
-import com.adobe.testing.s3mock.domain.S3Exception;
 import com.adobe.testing.s3mock.dto.ErrorResponse;
+import com.adobe.testing.s3mock.store.FileStore;
+import com.adobe.testing.s3mock.store.KmsKeyStore;
+import com.adobe.testing.s3mock.store.S3Exception;
 import java.util.ArrayList;
 import java.util.List;
 import javax.servlet.Filter;
