+++ server/src/main/java/com/adobe/testing/s3mock/store/AwsChunkedDecodingInputStream.java
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package com.adobe.testing.s3mock.domain;
+package com.adobe.testing.s3mock.store;
 
 import java.io.IOException;
 import java.io.InputStream;
