+++ server/src/main/java/com/adobe/testing/s3mock/store/DomainProperties.java
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package com.adobe.testing.s3mock.domain;
+package com.adobe.testing.s3mock.store;
 
 import java.util.HashSet;
 import java.util.List;
