+++ server/src/main/java/com/adobe/testing/s3mock/store/S3Object.java
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package com.adobe.testing.s3mock.domain;
+package com.adobe.testing.s3mock.store;
 
 import com.adobe.testing.s3mock.dto.Tag;
 import java.io.File;
