+++ server/src/main/java/com/adobe/testing/s3mock/util/MetadataUtil.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2017-2021 Adobe.
+ *  Copyright 2017-2022 Adobe.
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 
 package com.adobe.testing.s3mock.util;
 
-import com.adobe.testing.s3mock.domain.S3Object;
+import com.adobe.testing.s3mock.store.S3Object;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
