+++ server/src/test/java/com/adobe/testing/s3mock/store/KmsKeyStoreTest.java
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package com.adobe.testing.s3mock.domain;
+package com.adobe.testing.s3mock.store;
 
 import static org.assertj.core.api.Assertions.assertThat;
